Initial proof-of-concept implementation
This is a work-in-progress proof-of-concept. It doesn't compile
at the moment, but I wanted to get it out there.

The basic idea behind this is to automatically set up the services,
middlewares, and routes for routing RPC traffic to a range of different
providers, with block number informing the determination of the service's
health. At the moment, the block number subscription isn't working as
intended because gorilla websockets doesn't work with Yaegi. I'm working
on rewriting for a different websockets library, but wanted to make
current state available.